+++ tests/Integration/Message/Launch/Validator/Tool/ToolLaunchValidatorTest.php
@@ -116,8 +116,8 @@ class ToolLaunchValidatorTest extends TestCase
 
         $this->assertEquals(
             [
-                'ID token validation success',
                 'ID token kid header is provided',
+                'ID token validation success',
                 'ID token version claim is valid',
                 'ID token message_type claim is valid',
                 'ID token roles claim is valid',
@@ -157,8 +157,8 @@ class ToolLaunchValidatorTest extends TestCase
 
         $this->assertEquals(
             [
-                'ID token validation success',
                 'ID token kid header is provided',
+                'ID token validation success',
                 'ID token version claim is valid',
                 'ID token message_type claim is valid',
                 'ID token roles claim is valid',
@@ -205,8 +205,8 @@ class ToolLaunchValidatorTest extends TestCase
 
         $this->assertEquals(
             [
-                'ID token validation success',
                 'ID token kid header is provided',
+                'ID token validation success',
                 'ID token version claim is valid',
                 'ID token message_type claim is valid',
                 'ID token roles claim is valid',
@@ -482,6 +482,21 @@ class ToolLaunchValidatorTest extends TestCase
                     LtiMessagePayloadInterface::CLAIM_NONCE => 'existing',
                 ],
                 'ID token nonce claim already used'
+            ],
+            'Missing ID token deployment id' => [
+                [
+                    MessagePayloadInterface::HEADER_KID => $registration->getPlatformKeyChain()->getIdentifier()
+                ],
+                [
+                    MessagePayloadInterface::CLAIM_ISS => $registration->getPlatform()->getAudience(),
+                    MessagePayloadInterface::CLAIM_AUD => $registration->getClientId(),
+                    LtiMessagePayloadInterface::CLAIM_LTI_VERSION => LtiMessageInterface::LTI_VERSION,
+                    LtiMessagePayloadInterface::CLAIM_LTI_MESSAGE_TYPE => LtiMessageInterface::LTI_MESSAGE_TYPE_RESOURCE_LINK_REQUEST,
+                    LtiMessagePayloadInterface::CLAIM_LTI_ROLES => ['Learner'],
+                    LtiMessagePayloadInterface::CLAIM_SUB => 'user',
+                    LtiMessagePayloadInterface::CLAIM_NONCE => 'value',
+                ],
+                'ID token deployment_id claim is missing'
             ],
             'Invalid ID token deployment id' => [
                 [